refactor: 라벨 컴포넌트 색상 로직 수정
- 받아오는 문자열에 따라 색상이 바뀔 수 있도록 색상별 문자열 객체 추가
